Refactor replace inline event handlers with function calls
Inline event handlers can cause Syntax Errors. The reason is not strictly known,
but suspected to relate to how Fragment Parsing is performed in innerHTML. The
behaviour is consistent between at least Chrome and Firefox.

In this refactoring commit, the inline event handlers are extracted into functions,
and the inline code is replaced with a function call. This _should_ fix the problem
and at least all tests pass, so it shouldn't break anything unrelated.